# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Order Transaction API",
        "version": "1.0"
    },
    "host": "",
    "basePath": "/api/v1",
    "paths": {
...
        "/coupon": {
            "get": {
                "tags": [
                    "/coupon"
                ],
                "summary": "List Coupon",
                "description": "List coupon",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Coupon Code",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Coupon Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 10
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                }
            },
            "post": {
                "tags": [
                    "/coupon"
                ],
                "summary": "Create new Coupon",
                "description": "Create new Coupon",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "formData",
                        "description": "Coupon Code",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "valid_from",
                        "in": "formData",
                        "description": "Coupon Valid From (YYYY/MM/DD)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "valid_to",
                        "in": "formData",
                        "description": "Coupon Valid To (YYYY/MM/DD)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "coupon_value",
                        "in": "formData",
                        "description": "Coupon Value",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "coupon_percentage",
                        "in": "formData",
                        "description": "Coupon Percentage (in %)",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "formData",
                        "description": "Coupon limit (0 mean unlimited)",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/coupon/{id}": {
            "get": {
                "tags": [
                    "/coupon"
                ],
                "summary": "View Coupon detail",
                "description": "View Coupon detail",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Coupon ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/coupon"
                ],
                "summary": "Update existing Coupon",
                "description": "Update existing Coupon",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Coupon ID",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "valid_from",
                        "in": "formData",
                        "description": "Coupon Valid From (YYYY/MM/DD)",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "valid_to",
                        "in": "formData",
                        "description": "Coupon Valid To (YYYY/MM/DD)",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "coupon_value",
                        "in": "formData",
                        "description": "Coupon Value",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "coupon_percentage",
                        "in": "formData",
                        "description": "Coupon Percentage (in %)",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "formData",
                        "description": "Coupon limit (0 mean unlimited)",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/coupon"
                ],
                "summary": "Delete existing Coupon",
                "description": "Delete existing Coupon",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Coupon ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/coupon/{id}/status": {
            "get": {
                "tags": [
                    "/coupon"
                ],
                "summary": "Change Coupon Status",
                "description": "Change Coupon status",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Coupon ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/coupon/preview/{coupon_code}": {
            "get": {
                "tags": [
                    "/coupon"
                ],
                "summary": "Preview Coupon detail by product in cart",
                "description": "Preview Coupon detail by product in cart",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "coupon_code",
                        "in": "path",
                        "description": "Coupon Code",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "x-access-token",
            "in": "header"
        }
    },
...
}